+++ src/config/config.php
@@ -309,4 +309,9 @@ return [
     //
     'hideDefaultLocaleInURL' => false,
 
+    // If you want to display the locales in particular order in the language selector you should write the order here. 
+    //CAUTION: Please consider using the appropriate locale code otherwise it will not work
+    //Example: 'localesOrder' => ['es','en'],
+    'localesOrder' => [],
+
 ];