+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 1.2.3
+- Added `getLocalesOrder()` function to the package
+
+### 1.2
+- Added compatibility with Laravel 5.4
+
 ### 1.1
 - Added compatibility with Laravel 5.2